requirements/test.txt
flake8
mock == 2.0.0 ; python_version < '3.3'
pytest
pytest-django
